refactor : NameSplit utils 패키지지로 분리
