+++ src/test/java/racingcar/step2/utils/NameSplitTest.java
@@ -1,10 +1,8 @@
-package racingcar.step2.domain;
+package racingcar.step2.utils;
 
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
